% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bow.R
\name{bow}
\alias{bow}
\title{Bag of Words}
\usage{
bow(text, level = c("word", "letter"), sort = TRUE)
}
\arguments{
\item{text}{Text vector, whose elements can be phrases or documents. Note that both can be just one word}

\item{level}{It can be word or letter}

\item{sort}{Boolean value that lets the user define whether the order of the tokens will be based on most frequent or alphabetical order}
}
\value{
A data.frame, whose columns will be: (1) token, (2) frequency, (3) proportion
}
\description{
It builds a frequency table of tokens from a text or corpus, with frequencies of each token and proportion of the sample.
}
\examples{

# Example with a long text (Fisher, 1935: 6)
text1 <- "The discrepancy of opinion among historical writers on probability is so great that to mention the subject is unavoidable. It would, however, be out of place here to argue the point in detail. I will only state three considerations which will explain why, in the practical applications of the subject, I shall not assume the truth of Bayes’ axiom. Two of these reasons would, I think, be generally admitted, but the first, I can well imagine, might be indignantly repudiated in some quarters."
bow(text1)

# Example with multiple documents
text2 <- c("Bayes is from England", "Bayes is a statistician", "Bayes formulated Bayes Theorem")
bow(text2)

}
