% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2_co-ocurrence_test.R
\name{collocation_g2}
\alias{collocation_g2}
\title{G² lexical co-ocurrence test}
\usage{
collocation_g2(input, word1 = NULL, word2 = NULL, top_n = 20)
}
\arguments{
\item{input}{A data frame of ngrams (probably from ngrams()) or a text vector (no bag of words allowed here).}

\item{word1}{Optional. First word in a bigram.}

\item{word2}{Optional. Second word in a bigram.}

\item{top_n}{Integer. Size of the query of results.}
}
\value{
If word1 and word2 are given, a G² test result. Otherwise, a data frame of bigrams ranked by G² score (most significant collocations, that's what it is).
}
\description{
Applies the G² statistic for a given word pair, or all word pairs in a bigram list.
If word1 and word2 are not provided, it ranks all bigrams by significance (quite heavy to load sometimes, bear that in mind!)
}
