% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipf_heaps.R
\name{heaps}
\alias{heaps}
\title{Heaps' law calculation}
\usage{
heaps(text)
}
\arguments{
\item{text}{Text vector, whose elements can be phrases or documents. Note that both can be just a word}
}
\value{
data.frame with accumulated size, and quantity of unique words
}
\description{
Through this function we can estimate the growth of vocabulary with attention to the size of the corpus.
}
\examples{

# Example with a string (Euler, 1762)
text <- "there are persons determined to believe and admit nothing but what they see and touch"
heaps(text)

# Example with a dataset (heart_of_darkness, available in langstats, remember to load it with data(dataset_name))
data(heart_of_darkness)
heaps(heart_of_darkness)

}
