% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heart_of_darkness}
\alias{heart_of_darkness}
\title{Heart of Darkness}
\format{
A character vector.
}
\source{
Project Gutenberg
}
\usage{
heart_of_darkness
}
\description{
The complete text of the novel \emph{Heart of Darkness} by Joseph Conrad.
}
\keyword{datasets}
