% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handling.R
\name{input_handling}
\alias{input_handling}
\title{Input handling for functions of langstats: a smart and standard channel for inputs}
\usage{
input_handling(input, level = c("word", "letter"))
}
\arguments{
\item{input}{A text vector or a data frame, whether it comes from any function of langstats or any other origin}

\item{level}{It defines, as in many other functions here, the level of analysis required: it can be word or letter currently.}
}
\value{
A list, which is performance attractive, with two elements: (1) elements, that are linguistic elements, and (2) frequencies, which can be found or calculated; if both of these are not possible, the index will be NULL
}
\description{
Through input_handling we can extract tokens and frequencies from a variety of input formats. At the moment, it is sensitive to texts, and data frames, the former being a common output in this package, so this standarisation of inputs can create a really smooth workflow. It detects automatically tokens and frequencies from data frames, and in the case of texts, it tokenises and calculates frequencies, providing a bag of words.
}
\details{
When a data frame is provided, the function attempts to automatically detect a token column using common
names such as \code{"token"}, \code{"word"}, \code{"element"}, etc. If a frequency column named
\code{"frequency"} exists, it is normalized and returned. If only text is provided, the function tokenizes
using a regular expression based on the \code{level} argument, and completes the frequency index with a calculation on the frequencies of tokens. We plan to replace this search term procedure to a more functional one in next updates!
}
\examples{
text <- c("Codd is English", "Codd studied in Oxford", "Codd proposed the relational model in 1970")
input_handling(text, level = "word")

df <- data.frame(word = c("model", "relational", "Codd"), frequency = c(10, 3, 2))
input_handling(df, level = "word")

}
\keyword{data_handling}
