% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-grams.R
\name{ngrams}
\alias{ngrams}
\title{Generate n-grams from text or tokens}
\usage{
ngrams(input, n = 2, level = c("word", "letter"))
}
\arguments{
\item{input}{Text vector, whose elements can be phrases or documents, or data frame, for example, from the output of ngram(). Bear in mind that, as it has been defined, words must be on the first column! (Not anymore, input_handling() has acquired smart features!)}

\item{n}{Integer. The size of the n-grams to generate (e.g., 2 for bigrams). Just in case: it must be positive!}
}
\value{
A data frame with columns: (1) actual ngram, (2) integer vector of counts for each ngram, and (3) proportion of total n-grams.
}
}
\description{
This function generates sets of n-grams of the size n selected in the function. It supports text input, and data frame input, handled by the function input_handling().
}
\examples{
# Example with plain text (Cauchy, 1821: 1)
ngrams("In speaking of the continuity of functions, I could not dispense with a treatment of the principal properties of infinitely small quantities, properties which serve as the foundation of the infinitesimal calculus.", n = 2)

# Example with a data frame
df <- data.frame(word = c("Cauchy", "functions", "limits"), frequency = c(3, 10, 0))
input_handling(df, level = "word")

}
