% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{selected_piantadosi}
\alias{selected_piantadosi}
\title{Selected Papers by Steven Piantadosi}
\format{
A character vector.
}
\source{
\url{https://colala.berkeley.edu/people/piantadosi/}
}
\usage{
selected_piantadosi
}
\description{
A corpus of 5 academic texts by Steven T. Piantadosi, selected from the website. These are:
}
\keyword{datasets}
