% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{tf_idf}
\alias{tf_idf}
\title{TF-IDF: Term Frequency–Inverse Document Frequency}
\usage{
tf_idf(input, level = c("word", "letter"), token = c("regex", "transformer"))
}
\arguments{
\item{input}{A vector of character strings (one per document), or a data frame of tokens and their frequencies}

\item{level}{The level of linguistic granularity:}

\item{token}{The method for tokenisation}
}
\value{
A data frame with the following columns: (1) document, (2) term, (3) tf, (4) idf, and (5) tf_idf

texts <- c("Codd is English", "Codd studied in Oxford", "Codd proposed the relational model in 1970")
tf_idf(texts, level = "word")

df <- data.frame(word = c("model", "relational", "Codd"), frequency = c(10, 3, 2))
tf_idf(df, level = "word")

A data frame with columns: document, term, tf, idf, tf_idf
}
\description{
The function  calculates the TF-IDF score for each linguistic element across a set of documents. It combines two key components: term frequency (TF) and inverse document frequency (IDF), producing a score that reflects both the importance of a term within a document and its uniqueness across the full corpus.
}
