% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipf_heaps.R
\name{zipf}
\alias{zipf}
\title{Zipf's law calculation}
\usage{
zipf(text)
}
\arguments{
\item{text}{Text vector, whose elements can be phrases or documents. Note that both can be just a word}
}
\value{
data.frame Here the columns will be: (1) word, (2) frequency, (3) rank
}
\description{
This function grabs a text vector and returns a data frame with the words from the vector, and their frequencies and ranks.
}
\examples{
# Example with a string (Boyle, 1673, found in Fisher (1935))
 text <- "I am very sorry, Pyrophilus, that to the many (elsewhere enumerated) difficulties which you may meet with, and must therefore surmount, in the serious and effectual prosecution of experimental philosophy I must add one discouragement more, which will perhaps as much surprise as dishearten you; and it is, that besides that you will find (as we elsewhere mention) many of the experiments published by authors, or related to you by the persons you converse with, false and unsuccessful (besides this, I say), you will meet with several observations and experiments which, though communicated for true by candid authors or undistrusted eye-witnesses, or perhaps recommended by your own experience, may, upon further trial, disappoint your expectation, either not at all succeeding constantly, or at least varying much from what you expected."
zipf(text)

# Example with a dataset (selected_piantadosi, available in langstat, remember to load it with data(dataset_name))
data(selected_piantadosi)
zipf(selected_piantadosi)

}
