% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipf-mandelbrot.R
\name{zipf_mandelbrot}
\alias{zipf_mandelbrot}
\title{Mandelbrot's adjustment on Zipf's law}
\usage{
zipf_mandelbrot(text)
}
\arguments{
\item{text}{Text vector, whose elements can be phrases or documents. Note that both can be just one word}
}
\value{
List with the adjusted parameters, that is: alpha, beta, and C. Futhermore, original data are provided
}
\description{
Zipf-Mandelbrot's law it's a fix of the original Zipf's law, especifically designed to be more precise in low ranks (grammatical terms).
}
\examples{
# Example with multiple texts
text <- c("mandelbrot is polish", "mandelbrot is french", "mandelbrot is american")
zipf_mandelbrot(text)

# Example with a dataset (selected_piantadosi, available in langstats, remember to load it with data(dataset_name))
data(selected_piantadosi)
zipf_mandelbrot(selected_piantadosi)
}
