% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{idf}
\alias{idf}
\title{Inverse Document Frequency (IDF): measuring informational uniqueness}
\usage{
idf(input, level = c("word", "letter"), token = c("regex", "transformer"))
}
\arguments{
\item{input}{A text vector (multiple documents), or a data frame of tokens and frequencies}

\item{level}{The level of linguistic analysis:\link{TODO}}

\item{token}{The tokenisation method: either \link{TODO} (BERT-based)}
}
\value{
A data frame with the following columns: \link{TODO}
}
\description{
The function TODO calculates the inverse document frequency (IDF) for each linguistic element (token or character) across a collection of documents. The IDF metric helps to quantify how unique or informative a term is, relative to its distribution in the entire corpus.
}
\examples{
text <- c("Codd is English", "Codd studied in Oxford", "Codd proposed the relational model in 1970")
idf(text, level = "word")

df <- data.frame(word = c("model", "relational", "Codd"), frequency = c(10, 3, 2))
idf(df, level = "word")

}
